rag_system/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
motor==3.3.2
qdrant-client==1.7.3

# AI/ML
transformers==4.36.2
torch==2.1.2
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2

# Text Processing
nltk==3.8.1
spacy==3.7.2
tokenizers==0.15.0
tiktoken==0.5.2
beautifulsoup4==4.12.2

# Security
bcrypt==4.1.2
cryptography==41.0.8
passlib==1.7.4

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
aiofiles==23.2.1
httpx==0.26.0
loguru==0.7.2
python-dotenv==1.0.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
mypy==1.8.0
pre-commit==3.6.0